# server/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /assets/dishes/{filename}:
    get:
      operationId: StaticAssetsController_getDishImage
      parameters:
        - name: filename
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
...
  /auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User successfully registered
        '400':
          description: Bad request - validation failed
        '409':
          description: Email already exists
      summary: Register a new user
      tags: &ref_4
        - auth
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: User successfully logged in
        '401':
          description: Invalid credentials
      summary: Login user
      tags: *ref_4
  /auth/me:
    get:
      operationId: AuthController_getProfile
      parameters: []
      responses:
        '200':
          description: Returns current user profile
        '401':
          description: Unauthorized
      security:
        - JWT-auth: []
      summary: Get current user profile
      tags: *ref_4
...
  /dishes:
    post:
      operationId: DishesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDishDto'
      responses:
        '201':
          description: Dish created successfully
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
      security:
        - JWT-auth: []
      summary: Create new dish
      tags: &ref_6
        - dishes
    get:
      operationId: DishesController_findAll
      parameters:
        - name: includeDeleted
          required: false
          in: query
          description: Include deleted dishes
          schema:
            type: boolean
      responses:
        '200':
          description: Returns list of all dishes
      summary: Get all dishes
      tags: *ref_6
  /dishes/{id}:
    get:
      operationId: DishesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Dish ID
          schema:
            type: string
        - name: includeDeleted
          required: false
          in: query
          description: Include deleted dishes
          schema:
            type: boolean
      responses:
        '200':
          description: Returns dish details
        '404':
          description: Dish not found
      summary: Get dish by ID
      tags: *ref_6
    patch:
      operationId: DishesController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Dish ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDishDto'
      responses:
        '200':
          description: Dish updated successfully
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Dish not found
      security:
        - JWT-auth: []
      summary: Update dish
      tags: *ref_6
    delete:
      operationId: DishesController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Dish ID
          schema:
            type: string
      responses:
        '200':
          description: Dish deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Dish not found
      security:
        - JWT-auth: []
      summary: Delete dish (soft delete)
      tags: *ref_6
  /dishes/{id}/restore:
    post:
      operationId: DishesController_restore
      parameters:
        - name: id
          required: true
          in: path
          description: Dish ID
          schema:
            type: string
      responses:
        '200':
          description: Dish restored successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Dish not found
      security:
        - JWT-auth: []
      summary: Restore deleted dish
      tags: *ref_6
  /dishes/{id}/hard:
    delete:
      operationId: DishesController_hardDelete
      parameters:
        - name: id
          required: true
          in: path
          description: Dish ID
          schema:
            type: string
      responses:
        '200':
          description: Dish permanently deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Dish not found
      security:
        - JWT-auth: []
      summary: Permanently delete dish
      tags: *ref_6
...
  /menus/{id}/dishes/{dishId}:
    post:
      operationId: MenusController_addDishToMenu
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: dishId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags: *ref_7
    delete:
      operationId: MenusController_removeDishFromMenu
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: dishId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_7
...
  /categories:
    get:
      operationId: CategoriesController_findAll
      parameters: []
      responses:
        '200':
          description: Returns list of all categories
      summary: Get all categories
      tags: &ref_14
        - categories
    post:
      operationId: CategoriesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: Category created successfully
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
      security:
        - JWT-auth: []
      summary: Create new category
      tags: *ref_14
  /categories/{id}:
    get:
      operationId: CategoriesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Category ID
          schema:
            type: string
      responses:
        '200':
          description: Returns category details
        '404':
          description: Category not found
      summary: Get category by ID
      tags: *ref_14
    patch:
      operationId: CategoriesController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Category ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: Category updated successfully
        '400':
          description: Bad request - validation failed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Category not found
      security:
        - JWT-auth: []
      summary: Update category
      tags: *ref_14
    delete:
      operationId: CategoriesController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Category ID
          schema:
            type: string
      responses:
        '200':
          description: Category deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Category not found
      security:
        - JWT-auth: []
      summary: Delete category (soft delete)
      tags: *ref_14
  /categories/{id}/restore:
    post:
      operationId: CategoriesController_restore
      parameters:
        - name: id
          required: true
          in: path
          description: Category ID
          schema:
            type: string
      responses:
        '200':
          description: Category restored successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Category not found
      security:
        - JWT-auth: []
      summary: Restore deleted category
      tags: *ref_14
...
info:
  title: Restaurant Management API
  description: |2-

        # Restaurant Management System API Documentation
        
        Hệ thống quản lý nhà hàng toàn diện với các tính năng đầy đủ cho việc vận hành nhà hàng hiện đại.
        
        ## Tính năng chính
        - **🔐 Xác thực & Phân quyền**: JWT-based authentication với phân quyền theo vai trò
        - **👥 Quản lý người dùng**: Admin, Staff, Kitchen, Waiter, Warehouse roles
        - **🍽️ Quản lý thực đơn**: Categories, dishes, menus với theo dõi nguyên liệu
        - **📋 Quản lý đơn hàng**: Đặt bàn qua QR, workflow bếp, theo dõi trạng thái
        - **📦 Quản lý kho**: Theo dõi nguyên liệu, quản lý lô hàng, nhập/xuất kho
        - **💳 Thanh toán**: Tích hợp VNPay cho thanh toán online
        - **🪑 Quản lý bàn**: Tạo QR code, theo dõi trạng thái bàn
        - **📊 Báo cáo**: Báo cáo tài chính, báo cáo kho
        
        ## Xác thực API
        Hầu hết các endpoint yêu cầu xác thực qua JWT token. Thêm token vào header Authorization:
        `Authorization: Bearer <your-jwt-token>`
        
        ## Phân quyền người dùng
        - **ADMIN**: Toàn quyền hệ thống
        - **STAFF**: Các thao tác nhân viên phục vụ
        - **KITCHEN**: Các thao tác đặc thù bếp
        - **WAITER**: Quản lý bàn và đơn hàng
        - **WAREHOUSE**: Quản lý kho
        
        ## HTTP Status Codes
        API trả về các mã trạng thái HTTP chuẩn:
        - **200**: Thành công
        - **201**: Tạo mới thành công
...
servers:
...
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Nhập JWT token (không cần prefix Bearer)
      in: header
  schemas:
...
    RegisterDto:
      type: object
      properties:
        name:
          type: string
          description: User name
          example: Nguyễn Văn A
          minLength: 2
          maxLength: 50
        email:
          type: string
          description: User email address
          example: user@example.com
          format: email
        password:
          type: string
          description: User password
          example: password123
          minLength: 6
      required:
        - name
        - email
        - password
    LoginDto:
      type: object
      properties:
        email:
          type: string
          description: User email address
          example: user@example.com
          format: email
        password:
          type: string
          description: User password
          example: password123
          minLength: 6
      required:
        - email
        - password
...
    DishIngredientDto:
      type: object
      properties:
        ingredientId:
          type: string
          description: Ingredient ID
          example: 123e4567-e89b-12d3-a456-426614174000
        quantity:
          type: number
          description: Quantity of ingredient needed
          example: 100
          minimum: 0
      required:
        - ingredientId
        - quantity
    CreateDishDto:
      type: object
      properties:
        name:
          type: string
          description: Dish name
          example: Phở Bò Tái
        description:
          type: string
          description: Dish description
          example: Phở bò tái truyền thống với nước dùng đậm đà
        price:
          type: number
          description: Dish price in VND
          example: 85000
          minimum: 0
        image_url:
          type: string
          description: Image URL for the dish
          example: https://example.com/images/pho-bo-tai.jpg
        is_preparable:
          type: boolean
          description: Whether the dish can be prepared in kitchen
          example: true
        available:
          type: boolean
          description: Whether the dish is available for ordering
          example: true
        preparation_time:
          type: number
          description: Preparation time in minutes
          example: 15
          minimum: 0
        categoryId:
          type: string
          description: Category ID that this dish belongs to
          example: 123e4567-e89b-12d3-a456-426614174000
        ingredients:
          description: List of ingredients required for this dish
          type: array
          items:
            $ref: '#/components/schemas/DishIngredientDto'
      required:
        - name
        - description
        - price
        - preparation_time
        - categoryId
    UpdateDishDto:
      type: object
      properties: {}
...
    CreateCategoryDto:
      type: object
      properties:
        name:
          type: string
          description: Category name
          example: Món chính
          maxLength: 100
        description:
          type: string
          description: Category description
          example: Các món ăn chính trong thực đơn
      required:
        - name
    UpdateCategoryDto:
      type: object
      properties: {}
...