# server/src/dishes/dto/dish-ingredient.dto.ts
import { IsNotEmpty, IsNumber, IsUUID, Min } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class DishIngredientDto {
  @ApiProperty({
    description: 'Ingredient ID',
    example: '123e4567-e89b-12d3-a456-426614174000',
  })
  @IsNotEmpty()
  @IsUUID()
  ingredientId: string;
  
  @ApiProperty({
    description: 'Quantity of ingredient needed',
    example: 100,
    minimum: 0,
  })
  @IsNotEmpty()
  @IsNumber()
  @Min(0)
  quantity: number;
}